+++ MacForge/MacForge/Plugins/MF_Plugin.h
@@ -6,38 +6,40 @@
 //  Copyright © 2017 Wolfgang Baird. All rights reserved.
 //
 
+@import Paddle;
 #import <Foundation/Foundation.h>
 
-@interface MF_Plugin : NSObject
+@interface MF_Plugin : NSObject <PADProductDelegate>
 
 // Local or Repo
-@property NSString      *bundleID;  // Must be unique
-@property NSString      *bundleImage;
+@property NSString      *bundleID;              // Must be unique
+@property NSString      *bundleImage;           // Image to use for plugin
 
 // Repo Files
-@property NSDictionary  *webPlist;
-@property NSDictionary  *webPaddle;
-@property NSString      *webDeveloperDonate;
-@property NSString      *webDeveloperEmail;
-@property NSString      *webRepository;
-@property NSString      *webName;
-@property NSString      *webDescription;
-@property NSString      *webDescriptionShort;
-@property NSString      *webTarget;
-@property NSString      *webPublishDate;
-@property NSString      *webVersion;
-@property NSString      *webPrice;
-@property NSString      *webDeveloper;
-@property NSString      *webCompatability;
-@property NSString      *webFileName;
-@property NSString      *webSize;
-@property Boolean       webFeatured;
-@property Boolean       webPaid;
+@property NSDictionary  *webPlist;              // Copy of the plist info
+@property NSDictionary  *webPaddle;             // Paddle information
+@property NSString      *webDeveloperDonate;    // Donation url
+@property NSString      *webDeveloperEmail;     // Contact email
+@property NSString      *webRepository;         // Plugin repo                  -- Delete me
+@property NSString      *webName;               // Name of the plugin
+@property NSString      *webDescription;        // A longer description
+@property NSString      *webDescriptionShort;   // A short description
+@property NSString      *webTarget;             // Target applications
+@property NSString      *webPublishDate;        // Publish date
+@property NSString      *webVersion;            // Version
+@property NSString      *webPrice;              // Price
+@property NSString      *webDeveloper;          // Developer
+@property NSString      *webCompatability;      // macOS version compatability
+@property NSString      *webFileName;           // Download URL
+@property NSString      *webSize;               // Size of plugin
+@property Boolean       webFeatured;            // Plugin is featured
+@property Boolean       webPaid;                // Plugin is paid
+@property Boolean       webARM;                 // Plugin supports arm64/arm64e
 
 // Paddle
 @property NSString      *paddleEmail;
 @property NSString      *paddleLicense;
-//@property PADProduct    *paddleProduct;
+@property PADProduct    *paddleProduct;
 
 // Info
 @property Boolean           checkedPurchase;